chore: inline apollo-client and clean package.json
