+++ packages/author-profile/package.json
@@ -62,7 +62,6 @@
     "lodash.merge": "4.6.0",
     "prop-types": "15.5.10",
     "react": "16.0.0-alpha.6",
-    "react-apollo": "1.4.11",
     "react-dom": "15.5.4",
     "react-native-web": "0.0.97",
     "react-style-proptype": "3.0.0"
