+++ packages/provider/provider.stories.js
@@ -1,10 +1,16 @@
 import React from "react";
 import { Text } from "react-native";
 import { storiesOf } from "@storybook/react-native";
-import { ApolloProvider, gql } from "react-apollo";
+import {
+  ApolloClient,
+  ApolloProvider,
+  createNetworkInterface,
+  IntrospectionFragmentMatcher,
+  gql
+} from "react-apollo";
+
 import { withApolloProvider } from "storybook-addon-apollo-graphql";
 
-import client from "./fixtures/apollo-client";
 import connectGraphql, { AuthorProfileProvider } from "./provider.js";
 
 const Component = props => <Text>{JSON.stringify(props, null, 2)}</Text>;
@@ -27,6 +33,36 @@ storiesOf("Provider", module)
     const WithData = connectGraphql(query)(Component);
     return <WithData />;
   });
+
+const fragmentMatcher = new IntrospectionFragmentMatcher({
+  introspectionQueryResultData: {
+    __schema: {
+      types: [
+        {
+          kind: "UNION",
+          name: "Media",
+          possibleTypes: [
+            {
+              name: "Image"
+            },
+            {
+              name: "Video"
+            }
+          ]
+        }
+      ]
+    }
+  }
+});
+
+const networkInterface = createNetworkInterface({
+  uri: "http://localhost:4000/graphql/"
+});
+
+const client = new ApolloClient({
+  networkInterface,
+  fragmentMatcher
+});
 
 storiesOf("Provider", module).add("AuthorProfileProvider", () =>
   <ApolloProvider client={client}>
